Remove Bjy Authorize and place test to common folder
Bjy Authorize has been removed because it force the definition of
stateful assertions reducing the possibility to cache these classes
and making testing difficult.
Zend Framework Acl pass as RoleInterface parameter in assert method
of an Assertion the current user so every assertion has been converted
into stateless class.

Test folder has been renamed "tests" because it seems widely used
by php project this convension, even if the folder name is plural
and the others folders are singular.
