+++ composer.json
@@ -19,8 +19,7 @@
 		"prooph/event-sourcing": "~1.0",
 		"prooph/event-store": "~2.0",
 		"prooph/prooph-event-store-module": "~2.0",
-		"rwoverdijk/assetmanager": "*",
-		"bjyoungblood/bjy-authorize": "1.4.*"
+		"rwoverdijk/assetmanager": "*"
 	},
 	"require-dev": {
 		"phpunit/phpunit": ">=4.1",
