+++ config/autoload/auth.unit.php
@@ -1,0 +1,11 @@
+<?php
+
+use ZFX\Authentication\AuthenticationServiceMock;
+
+return array(
+	'service_manager' => array(
+		'invokables' => array(
+			'Zend\Authentication\AuthenticationService' => AuthenticationServiceMock::class
+		),
+	),
+);
