--- config/autoload/bjyauthorize.global.php
@@ -1,16 +1,0 @@
-<?php 
-use Application\Entity\User;
-
-return array(
-
-	'bjyauthorize' => array( 
-		 
-		'identity_provider' => 'Authorization\CurrentUserProvider',
-
-		'role_providers' => array(
-			'BjyAuthorize\Provider\Role\Config' => User::getRoleCollection(), 
-		),
-
-	),	
-)
-?>
