+++ module/Application/src/Application/Entity/User.php
@@ -6,14 +6,13 @@ use Zend\Permissions\Acl\Role\RoleInterface;
 use Doctrine\ORM\Mapping AS ORM;
 use Doctrine\Common\Collections\ArrayCollection;
 use Rhumsaa\Uuid\Uuid;
-use BjyAuthorize\Provider\Identity\ProviderInterface;	
 
 
 /**
  * @ORM\Entity @ORM\Table(name="users")
  *
  */
-class User implements ProviderInterface, RoleInterface
+class User implements RoleInterface
 {	   
 	CONST STATUS_ACTIVE = 1;
 	CONST ROLE_ADMIN = 'admin';
@@ -22,15 +21,6 @@ class User implements ProviderInterface, RoleInterface
 	
 	CONST EVENT_CREATED = "User.Created";
 
-	private static $roles=[ 
-		self::ROLE_GUEST => [], 
-		self::ROLE_USER => [
-			'children' => [
-				self::ROLE_ADMIN => [],
-			 ], 
-		]
-	];
-	
 	/**
 	 * @ORM\Id @ORM\Column(type="string") 
 	 * @var string
@@ -118,8 +108,8 @@ class User implements ProviderInterface, RoleInterface
 		return $rv;
 	}
 	
-	public function addOrganizationMembership($membership){
-		$this->memberships->add($membership);
+	public function addOrganizationMembership(OrganizationMembership $membership){
+		$this->memberships->set($membership->getOrganization()->getId(), $membership);
 	}
 	
 	public function getId() {
@@ -232,27 +222,23 @@ class User implements ProviderInterface, RoleInterface
 	/**
 	 * 
 	 * @param id|Organization $organization
-	 * @return unknown
+	 * @return bool
 	 */
 	public function isMemberOf($organization) {
 		$key = $organization instanceof Organization ? $organization->getId() : $organization;
 		return $this->memberships->containsKey($key);
 	}
 
-	public function getIdentityRoles(){
-		return $this->role;
-	}
-
 	public function setRole($role){
 		$this->role = $role;
+		return $this;
 	}
 	
-	public static function getRoleCollection(){
-		return self::$roles;
-
+	public function getRole() {
+		return $this->role;
 	}
 	
 	public function getRoleId(){
-		return $this->getIdentityRoles();
+		return $this->getRole();
 	}
 }
