+++ module/TaskManagement/src/TaskManagement/Assertion/NotCompletedTaskAssertion.php
@@ -10,8 +10,8 @@ use TaskManagement\Entity\Task;
 
 class NotCompletedTaskAssertion implements AssertionInterface{
 	
-	public function assert(Acl $acl, RoleInterface $user = null, ResourceInterface $resource = null, $privilege = null){
-		
-		return $resource->getStatus() < Task::STATUS_COMPLETED;		
+	public function assert(Acl $acl, RoleInterface $user = null, ResourceInterface $resource = null, $privilege = null)
+	{
+		return $resource->getStatus() < Task::STATUS_COMPLETED;
 	}
 }
