+++ module/TaskManagement/src/TaskManagement/Assertion/TaskMemberNotOwnerAndNotCompletedTaskAssertion.php
@@ -2,37 +2,16 @@
 
 namespace TaskManagement\Assertion;
 
-
 use Zend\Permissions\Acl\Acl;
 use Zend\Permissions\Acl\Resource\ResourceInterface;
 use Zend\Permissions\Acl\Role\RoleInterface;
-use Application\Entity\User;
 use TaskManagement\Entity\Task;
 
 class TaskMemberNotOwnerAndNotCompletedTaskAssertion extends NotCompletedTaskAssertion
 {
-    private $loggedUser;
-    
-	public function setLoggedUser($loggedUser = null) {
-    	$this->loggedUser = $loggedUser;
-    }
-    
-	public function assert(Acl $acl, RoleInterface $role = null, ResourceInterface $resource = null, $privilege = null){
-		
-		if(parent::assert($acl, $role, $resource, $privilege)){
-			
-			if($this->loggedUser instanceof User){
-
-				if($resource->hasMember($this->loggedUser)){
-
-					$roleMember = $resource->getMemberRole($this->loggedUser);
-					if($roleMember != Task::ROLE_OWNER){
-						return true;
-					}		
-				}
-			}
-		}
-		
-		return false;
-    }    
+	public function assert(Acl $acl, RoleInterface $user = null, ResourceInterface $resource = null, $privilege = null)
+	{
+		return parent::assert($acl, $user, $resource, $privilege)
+			&& $resource->getMemberRole($user != Task::ROLE_OWNER);
+	}
 }
