+++ module/TaskManagement/src/TaskManagement/Assertion/TaskOwnerAndCompletedTaskWithEstimationProcessCompletedAssertion.php
@@ -2,42 +2,18 @@
 
 namespace TaskManagement\Assertion;
 
-
 use Zend\Permissions\Acl\Acl;
-use Zend\Permissions\Acl\Resource\ResourceInterface;
 use Zend\Permissions\Acl\Role\RoleInterface;
+use Zend\Permissions\Acl\Resource\ResourceInterface;
 use Zend\Permissions\Acl\Assertion\AssertionInterface;
-use Application\Entity\User;
 use TaskManagement\Entity\Task;
 
 class TaskOwnerAndCompletedTaskWithEstimationProcessCompletedAssertion implements AssertionInterface
 {
-    private $loggedUser;
-    
-	public function setLoggedUser($loggedUser = null) {
-    	$this->loggedUser = $loggedUser;
-    }
-    
-	public function assert(Acl $acl, RoleInterface $role = null, ResourceInterface $resource = null, $privilege = null){
-		
-		if($resource->getStatus() == Task::STATUS_COMPLETED) {
-			
-			if($this->loggedUser instanceof User){
-
-				if(!is_null($resource->getAverageEstimation())){
-				
-					if($resource->hasMember($this->loggedUser)){
-						
-						$roleMember = $resource->getMemberRole($this->loggedUser->getId());
-						if($roleMember == Task::ROLE_OWNER){
-							
-							return true;
-						}						
-					}					
-				}				
-    		}
-    	}    	
-    	return false;
-    }
-    
+	public function assert(Acl $acl, RoleInterface $user = null, ResourceInterface $resource = null, $privilege = null)
+	{
+		return $resource->getStatus() == Task::STATUS_COMPLETED
+			&& $resource->getAverageEstimation() != null
+			&& $resource->getMemberRole($user) == Task::ROLE_OWNER;
+	}
 }
