+++ module/ZFX/Module.php
@@ -1,0 +1,16 @@
+<?php
+namespace ZFX;
+
+class Module
+{
+	public function getAutoloaderConfig()
+	{		
+		return array(
+			'Zend\Loader\StandardAutoloader' => array(
+				'namespaces' => array(
+					__NAMESPACE__ => __DIR__ . '/src/' . __NAMESPACE__,
+				),
+			),
+		);
+	}
+}
