+++ module/ZFX/src/ZFX/Controller/Plugin/IsAllowed.php
@@ -1,0 +1,36 @@
+<?php
+namespace ZFX\Controller\Plugin;
+
+use Zend\Mvc\Controller\Plugin\AbstractPlugin;
+use Zend\Permissions\Acl\Acl;
+
+/**
+ * IsAllowed Controller plugin. Allows checking access to a resource/privilege in controllers.
+ *
+ */
+class IsAllowed extends AbstractPlugin
+{
+	/**
+	 * @var Acl
+	 */
+	protected $acl;
+
+	/**
+	 * @param Acl $acl
+	 */
+	public function __construct(Acl $acl)
+	{
+		$this->acl = $acl;
+	}
+
+	/**
+	 * @param mixed	  $resource
+	 * @param mixed|null $privilege
+	 *
+	 * @return bool
+	 */
+	public function __invoke($user, $resource, $privilege = null)
+	{
+		return $this->acl->isAllowed($user, $resource, $privilege);
+	}
+}
